Allow manually setting the UID for grafana
This code change is to resolve permission issues with host volumes being mapped to the container. As grafana is not run as root and the user "grafana" created within the container gets a random UID, this code help manually set the UID/GID for grafana to meet the host users needs. 

If environment variables are not provided, the code continues as-is, but if a UID="1000" and GID="1000" is provided as environment variables during container initiation, grafana's uid/gid are changed. In a scenario where a host volume is mapped to GF_PATHS_DATA, chown-ing with the UID of the container initiating user itself will lead to less of a permissions headache.