Tag master builds of grafana with the changeset.
This PR changes the behaviour of repo a fair bit; firstly, a GRAFANA_VERSION must be specified, otherwise the build will be a no-op.  Then, if the GRAFANA_VERSION begins with the character v, we will package a full release; otherwise we will download and package what ever GRAFANA_VERSION is set to and push it to grafana/grafana-dev.

Also, trigger a deploy to k8s after all this.

Signed-off-by: Tom Wilkie <tom.wilkie@gmail.com>